petcare/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
Js2Py==0.71
packaging==21.3
pipwin==0.5.2
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyGreSQL==5.2.4
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.16.4
whitenoise==6.2.0
